# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cabo Fit Pass API
  version: 1.0.0
  description: |
    **Cabo Fit Pass** is a production-ready fitness booking platform built for the Los Cabos market. 
    
    This API provides a comprehensive suite of endpoints for:
    - **User Authentication & Profiles** - Secure user management with NextAuth.js
    - **Class Booking System** - Credit-based booking with atomic transactions
    - **Payment Processing** - Stripe integration with webhook support
    - **Admin Management** - Gym and studio administration
    - **Health Monitoring** - System health and performance tracking
    
    ## Authentication
    The API uses JWT-based authentication via NextAuth.js. Include the session token in requests to protected endpoints.
    
    ## Rate Limiting
    API endpoints are rate-limited to prevent abuse:
    - **Standard endpoints**: 100 requests per minute per IP
    - **Payment endpoints**: 20 requests per minute per user
    - **Admin endpoints**: 50 requests per minute per admin user
    
    ## Error Handling
    All endpoints return consistent error responses with:
    - HTTP status codes following REST conventions
    - Structured error objects with `success: false`
    - Detailed error messages for debugging (development only)
    - Request IDs for tracing (via headers)

  contact:
    name: Cabo Fit Pass API Support
    url: https://github.com/your-repo/cabo-fit-pass
    email: support@cabofitpass.com
  license:
    name: MIT
...
servers:
...
security:
  - BearerAuth: []
  - SessionAuth: []
paths:
...
  /profile:
    get:
      tags:
        - Profile
      summary: Get current user profile
      description: |
        Retrieves the authenticated user's profile including:
        - Basic user information
        - Credit balance
        - Subscription tier
        - Account preferences
        
        **Caching**: Results are cached for 1 hour for performance
      operationId: getUserProfile
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          profile:
                            $ref: '#/components/schemas/UserProfile'
              example:
                success: true
                data:
                  profile:
                    id: "user-123"
                    email: "john@example.com"
                    full_name: "John Doe"
                    user_type: "member"
                    credits: 25
                    subscription_tier: "premium"
                    created_at: "2024-01-15T08:00:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Profile
      summary: Update user profile
      description: |
        Updates user profile information. Only modifiable fields are accepted.
        
        **Rate Limiting**: 10 updates per minute per user
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                  maxLength: 100
                  description: User's full display name
                subscription_tier:
                  type: string
                  enum: [free, basic, premium, unlimited]
                  description: User's subscription level
            example:
              full_name: "John Smith"
              subscription_tier: "premium"
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          profile:
                            $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Classes
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from authentication
    
    SessionAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth session cookie
  schemas:
...
    UserProfile:
      type: object
      required:
        - id
        - email
        - user_type
        - credits
        - subscription_tier
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        full_name:
          type: string
          nullable: true
          maxLength: 100
          description: User's full display name
        user_type:
          type: string
          enum: [member, instructor, studio_owner]
          description: User's role in the system
        credits:
          type: integer
          minimum: 0
          description: Current credit balance
        subscription_tier:
          type: string
          enum: [free, basic, premium, unlimited]
          description: User's subscription level
        created_at:
          type: string
          format: date-time
          nullable: true
          description: Account creation timestamp
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: Last profile update timestamp
      example:
        id: "user-123"
        email: "john@example.com"
        full_name: "John Doe"
        user_type: "member"
        credits: 25
        subscription_tier: "premium"
        created_at: "2024-01-15T08:00:00.000Z"
        updated_at: "2024-01-27T10:30:00.000Z"
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          nullable: true
          description: Response data (varies by endpoint)
        error:
          type: string
          nullable: true
          description: Error message if success is false
        message:
          type: string
          nullable: true
          description: Additional information or success message
      example:
        success: true
        message: "Operation completed successfully"
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
          description: Error message
        code:
          type: string
          nullable: true
          description: Error code for programmatic handling
        details:
          type: object
          nullable: true
          description: Additional error details
      example:
        success: false
        error: "Validation failed"
        code: "VALIDATION_ERROR"
        details:
          field: "email"
          message: "Invalid email format"
  responses:
    BadRequest:
      description: Bad Request - Invalid input or request format
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Invalid request parameters"
            code: "BAD_REQUEST"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required"
            code: "UNAUTHORIZED"
...
    InternalServerError:
      description: Internal Server Error - Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Internal server error"
            code: "INTERNAL_ERROR"
...